requirements.txt
pandas==1.1.4
reportlab==3.5.55
validate_email==1.3
Pillow==8.1.0
XlsxWriter==1.3.7
xlrd==1.2.0
